+++ taguette/web.py
@@ -162,6 +162,8 @@ class Login(BaseHandler):
                         next=self.get_argument('next', ''))
 
     def post(self):
+        if not self.application.multiuser:
+            raise HTTPError(404)
         login = self.get_body_argument('login')
         password = self.get_body_argument('password')
         user = self.db.query(database.User).get(login)
@@ -200,6 +202,8 @@ class Register(BaseHandler):
             self.render('login.html', register=True)
 
     def post(self):
+        if not self.application.multiuser:
+            raise HTTPError(404)
         if not self.application.register_enabled:
             raise HTTPError(403)
         login = self.get_body_argument('login')
@@ -303,7 +307,6 @@ class ProjectMeta(BaseHandler):
         project.name = obj['name']
         project.description = obj['description']
         now = datetime.utcnow()
-        project.meta_updated = now
         logger.info("Updated project: %r %r",
                     project.name, project.description)
         cmd = database.Command.project_meta(
@@ -346,7 +349,7 @@ class DocumentAdd(BaseHandler):
                 contents=body,
             )
             self.db.add(doc)
-            project.documents_updated = datetime.utcnow()
+            self.db.flush()  # Need to flush to get doc.id
             cmd = database.Command.document_add(
                 self.current_user,
                 doc,
@@ -358,6 +361,44 @@ class DocumentAdd(BaseHandler):
             self.db.refresh(cmd)
             self.application.notify_project(project.id, cmd)
             self.send_json({'created': doc.id})
+
+
+class DocumentUpdate(BaseHandler):
+    @authenticated
+    def post(self, project_id, document_id):
+        obj = self.get_json()
+        document = self.get_document(project_id, document_id)
+        if obj:
+            if 'name' in obj:
+                document.name = obj['name']
+            if 'description' in obj:
+                document.description = obj['description']
+            cmd = database.Command.document_add(
+                self.current_user,
+                document,
+            )
+            self.db.add(cmd)
+            self.db.commit()
+            self.db.refresh(cmd)
+            self.application.notify_project(document.project_id, cmd)
+
+        self.send_json({'id': document.id})
+
+    @authenticated
+    def delete(self, project_id, document_id):
+        document = self.get_document(project_id, document_id)
+        self.db.delete(document)
+        cmd = database.Command.document_delete(
+            self.current_user,
+            document,
+        )
+        self.db.add(cmd)
+        self.db.commit()
+        self.db.refresh(cmd)
+        self.application.notify_project(document.project_id, cmd)
+
+        self.set_status(204)
+        self.finish()
 
 
 class DocumentContents(BaseHandler):
@@ -390,7 +431,7 @@ class HighlightAdd(BaseHandler):
                                 end_offset=end,
                                 snippet=snippet)
         self.db.add(hl)
-        self.db.commit()  # Need to commit to get hl.id
+        self.db.flush()  # Need to flush to get hl.id
         self.db.bulk_insert_mappings(database.HighlightTag, [
             dict(
                 highlight_id=hl.id,
@@ -420,14 +461,7 @@ class HighlightUpdate(BaseHandler):
         hl = self.db.query(database.Highlight).get(int(highlight_id))
         if hl.document_id != document.id:
             raise HTTPError(404)
-        if not obj:
-            self.db.delete(hl)
-            cmd = database.Command.highlight_delete(
-                self.current_user,
-                document,
-                hl.id,
-            )
-        else:
+        if obj:
             if 'start_offset' in obj:
                 hl.start_offset = obj['start_offset']
             if 'end_offset' in obj:
@@ -450,12 +484,32 @@ class HighlightUpdate(BaseHandler):
                 hl,
                 obj.get('tags', []),
             )
+            self.db.add(cmd)
+            self.db.commit()
+            self.db.refresh(cmd)
+            self.application.notify_project(document.project_id, cmd)
+
+        self.send_json({'id': hl.id})
+
+    @authenticated
+    def delete(self, project_id, document_id, highlight_id):
+        document = self.get_document(project_id, document_id)
+        hl = self.db.query(database.Highlight).get(int(highlight_id))
+        if hl.document_id != document.id:
+            raise HTTPError(404)
+        self.db.delete(hl)
+        cmd = database.Command.highlight_delete(
+            self.current_user,
+            document,
+            hl.id,
+        )
         self.db.add(cmd)
         self.db.commit()
         self.db.refresh(cmd)
         self.application.notify_project(document.project_id, cmd)
 
-        self.send_json({'id': hl.id})
+        self.set_status(204)
+        self.finish()
 
 
 class Highlights(BaseHandler):
@@ -493,7 +547,7 @@ class TagAdd(BaseHandler):
         tag = database.Tag(project=project,
                            path=obj['path'], description=obj['description'])
         self.db.add(tag)
-        self.db.commit()  # Need to commit to get tag.id
+        self.db.flush()  # Need to flush to get tag.id
         cmd = database.Command.tag_add(
             self.current_user,
             tag,
@@ -514,14 +568,7 @@ class TagUpdate(BaseHandler):
         tag = self.db.query(database.Tag).get(int(tag_id))
         if tag.project_id != project.id:
             raise HTTPError(404)
-        if not obj:
-            self.db.delete(tag)
-            cmd = database.Command.tag_delete(
-                self.current_user,
-                project.id,
-                tag.id,
-            )
-        else:
+        if obj:
             if 'path' in obj:
                 tag.path = obj['path']
             if 'description' in obj:
@@ -530,12 +577,32 @@ class TagUpdate(BaseHandler):
                 self.current_user,
                 tag,
             )
+            self.db.add(cmd)
+            self.db.commit()
+            self.db.refresh(cmd)
+            self.application.notify_project(project.id, cmd)
+
+        self.send_json({'id': tag.id})
+
+    @authenticated
+    def delete(self, project_id, tag_id):
+        project = self.get_project(project_id)
+        tag = self.db.query(database.Tag).get(int(tag_id))
+        if tag.project_id != project.id:
+            raise HTTPError(404)
+        self.db.delete(tag)
+        cmd = database.Command.tag_delete(
+            self.current_user,
+            project.id,
+            tag.id,
+        )
         self.db.add(cmd)
         self.db.commit()
         self.db.refresh(cmd)
         self.application.notify_project(project.id, cmd)
 
-        self.send_json({'id': tag.id})
+        self.set_status(204)
+        self.finish()
 
 
 class ProjectEvents(BaseHandler):
@@ -569,6 +636,8 @@ class ProjectEvents(BaseHandler):
         elif type_ == 'document_add':
             payload['id'] = cmd.document_id
             result = {'document_add': [payload]}
+        elif type_ == 'document_delete':
+            result = {'document_delete': [cmd.document_id]}
         elif type_ == 'highlight_add':
             result = {'highlight_add': {cmd.document_id: [payload]}}
         elif type_ == 'highlight_delete':
@@ -733,32 +802,41 @@ def make_app(db_url, multiuser, register_enabled=True, debug=False):
 
     return Application(
         [
+            # Basic pages
             URLSpec('/', Index, name='index'),
             URLSpec('/login', Login, name='login'),
             URLSpec('/logout', Logout, name='logout'),
             URLSpec('/register', Register, name='register'),
             URLSpec('/new', NewProject, name='new_project'),
+
+            # Project view
             URLSpec('/project/([0-9]+)', Project, name='project'),
             URLSpec('/project/([0-9]+)/document/[0-9]+', Project,
                     name='project_doc'),
             URLSpec('/project/([0-9]+)/highlights/[^/]+', Project,
                     name='project_tag'),
-            URLSpec('/project/([0-9]+)/meta', ProjectMeta),
-            URLSpec('/project/([0-9]+)/document/new', DocumentAdd),
-            URLSpec('/project/([0-9]+)/document/([0-9]+)/content',
-                    DocumentContents),
-            URLSpec('/project/([0-9]+)/document/([0-9]+)/highlight/new',
-                    HighlightAdd),
-            URLSpec('/project/([0-9]+)/document/([0-9]+)/highlight/([0-9]+)',
-                    HighlightUpdate),
-            URLSpec('/project/([0-9]+)/highlights/([^/]+)/list', Highlights),
-            URLSpec('/project/([0-9]+)/tag/new', TagAdd),
-            URLSpec('/project/([0-9]+)/tag/([0-9]+)', TagUpdate),
-            URLSpec('/project/([0-9]+)/events', ProjectEvents),
+
+            # Export options
             URLSpec('/project/([0-9]+)/export/codebook.csv', ExportCodebookCsv,
                     name='export_codebook_csv'),
             URLSpec('/project/([0-9]+)/export/codebook.docx',
                     ExportCodebookDoc, name='export_codebook_doc'),
+
+            # API
+            URLSpec('/api/project/([0-9]+)', ProjectMeta),
+            URLSpec('/api/project/([0-9]+)/document/new', DocumentAdd),
+            URLSpec('/api/project/([0-9]+)/document/([0-9]+)', DocumentUpdate),
+            URLSpec('/api/project/([0-9]+)/document/([0-9]+)/content',
+                    DocumentContents),
+            URLSpec('/api/project/([0-9]+)/document/([0-9]+)/highlight/new',
+                    HighlightAdd),
+            URLSpec(
+                '/api/project/([0-9]+)/document/([0-9]+)/highlight/([0-9]+)',
+                HighlightUpdate),
+            URLSpec('/api/project/([0-9]+)/highlights/([^/]+)', Highlights),
+            URLSpec('/api/project/([0-9]+)/tag/new', TagAdd),
+            URLSpec('/api/project/([0-9]+)/tag/([0-9]+)', TagUpdate),
+            URLSpec('/api/project/([0-9]+)/events', ProjectEvents),
         ],
         static_path=pkg_resources.resource_filename('taguette', 'static'),
         login_url='/login',